Start using proper address types in multiboot2 stuff and get rid of virtual_base hack

+++ kernel/x86_64/src/memory/stack_allocator.rs
@@ -83,7 +83,7 @@ impl StackAllocator
                     active_table.map(page, paging::entry::EntryFlags::WRITABLE, frame_allocator);
                 }
 
-                let top_of_stack = end.get_start_address().offset(PAGE_SIZE);
+                let top_of_stack = end.get_start_address().offset(PAGE_SIZE as isize);
                 Some(Stack::new(top_of_stack, start.get_start_address()))
             }
